removed node hover event
